{{cookiecutter.prj_name}}/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
click==7.1.2
Flask==1.1.4
flask-accepts==0.18.2
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
flask-restx==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
Werkzeug==1.0.1
